requirements.txt
# Core Django
Django==4.2.9
djangorestframework==3.14.0
django-filter==23.5

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Production server
gunicorn==21.2.0

# CORS headers for API
django-cors-headers==4.3.1

# Environment variables
python-decouple==3.8

# Excel/CSV export
openpyxl==3.1.2
xlsxwriter==3.1.9

# Timezone support
pytz==2023.3

# Future SSO support (commented for initial release)
# django-allauth==0.57.0
# social-auth-app-django==5.4.0
